Update lg_cover.html
This is my first attempt to edit, so wanted this to be a limit change. I assume that the best practice for later changes is to make all/numerous changes, the propose. For this trial run, I want to get rid of the docket number note entirely, so I deleted lines 21 - 34. I assume that for instances where I am modify the text of an existing note (probably the most common type of change), I will simply revise the existing "visible" text, eg, lines 17-18 of the "CV Plain Language" note. My thinking for inserting a new note will be to copy, insert and modify the lines from an appropriate similar existing note. Let me know if I'm doing something wrong.